requirements.txt
pandas==3.0.2
PyYAML>=6.0
pydantic>=2.0
pydantic-settings>=2.0
numpy>=1.26
fastapi>=0.111
uvicorn[standard]>=0.29
asyncpg>=0.29
httpx>=0.27
starlette>=0.37
